@@ -8,10 +8,10 @@ import BatteryCore
 class BatteryViewModel: ObservableObject {
     // Real battery info
     @Published var batteryInfo = BatteryInfo()
-    
+
     // Error state for user feedback
     @Published var errorMessage: String? = nil
-    
+
     // App settings
     @Published var launchAtStartup = LaunchAtStartup.isEnabled {
         didSet {
@@ -31,7 +31,7 @@ class BatteryViewModel: ObservableObject {
             startMonitoring()
         }
     }
-    
+
     // Widget settings
     @Published var showDesktopWidget = false {
         didSet {
@@ -53,7 +53,7 @@ class BatteryViewModel: ObservableObject {
             }
         }
     }
-    
+
     // Auto-update settings
     @Published var checkForUpdatesAutomatically = false {
         didSet {
@@ -66,89 +66,116 @@ class BatteryViewModel: ObservableObject {
             }
         }
     }
-    
-    
+
+    // Notch glow alert settings
+    @Published var enableNotchAlerts = true {
+        didSet {
+            saveSetting("enableNotchAlerts", value: enableNotchAlerts)
+        }
+    }
+
+    // Startup notch glow animation
+    @Published var enableNotchStartupAnimation = true {
+        didSet {
+            saveSetting("enableNotchStartupAnimation", value: enableNotchStartupAnimation)
+        }
+    }
+
     private let reader = BatteryReader()
-    private var timer: Timer?
+    private var batteryRefreshTask: Task<Void, Never>?
+    private var updateCheckTask: Task<Void, Never>?
     private var cancellables = Set<AnyCancellable>()
     private let logger = Logger(subsystem: "com.microverse.app", category: "BatteryViewModel")
     private var widgetManager: DesktopWidgetManager?
-    private var updateCheckTimer: Timer?
-    
+
+    // Track previous state for alert transitions
+    private var previousBatteryCharge: Int = -1
+    private var previousIsCharging: Bool = false
+    private var previousIsPluggedIn: Bool = false
+    private var hasShownLowBatteryAlert = false
+    private var hasShownCriticalBatteryAlert = false
+    private var hasPlayedStartupNotchAnimation = false
+
     // Enhanced notch system
     private let notchViewModel: MicroverseNotchViewModel
-    
+
     init() {
         logger.info("BatteryViewModel initializing...")
-        
+
         // Initialize notch system with proper dependency injection
         notchViewModel = MicroverseNotchViewModel()
-        
+
         loadSettings()
         refreshBatteryInfo()
         startMonitoring()
         setupBindings()
-        
+
         // Initialize widget manager
         widgetManager = DesktopWidgetManager(viewModel: self)
-        
+
         // Initialize enhanced notch system
         notchViewModel.setBatteryViewModel(self)
         NotchServiceLocator.register(notchViewModel)
-        
-        // Observe notch layout mode changes to trigger UI updates
+
+        // Ensure Settings UI updates when notch mode changes externally (e.g., context menu)
         notchViewModel.$layoutMode
             .sink { [weak self] _ in
-                // Trigger UI update by changing @Published property
                 self?.objectWillChange.send()
             }
             .store(in: &cancellables)
-        
-        // Show notch if layout mode is not off (after settings are loaded)
-        if notchViewModel.layoutMode != .off {
+
+        // Show notch on startup if enabled and supported
+        if notchViewModel.layoutMode != .off, isNotchAvailable {
             Task { @MainActor in
                 do {
                     try await notchViewModel.showNotch()
                     logger.info("Notch displayed on startup with mode: \(self.notchViewModel.layoutMode.displayName)")
+                    await playStartupNotchAnimationIfNeeded()
                 } catch {
                     logger.error("Failed to show notch on startup: \(error.localizedDescription)")
                 }
             }
+        } else {
+            // If the notch UI isn't enabled, we may still want a startup glow (physical notch fallback).
+            Task { @MainActor [weak self] in
+                guard let self else { return }
+                try? await Task.sleep(nanoseconds: 900_000_000)
+                await self.playStartupNotchAnimationIfNeeded()
+            }
         }
-        
-        // Show widget if it was enabled (adaptive display service will manage this)
+
+        // Show widget if it was enabled
         if showDesktopWidget {
             widgetManager?.showWidget()
         }
-        
+
         // Start automatic update checking if enabled
         if checkForUpdatesAutomatically {
             schedulePeriodicUpdateCheck()
         }
-        
+
         logger.info("BatteryViewModel initialized")
     }
-    
+
     deinit {
-        timer?.invalidate()
-        updateCheckTimer?.invalidate()
-        
-        // Manual cleanup - schedule on main actor
+        batteryRefreshTask?.cancel()
+        updateCheckTask?.cancel()
+
         Task { @MainActor in
             NotchServiceLocator.unregister()
         }
         logger.info("BatteryViewModel deallocated")
     }
-    
+
     // MARK: - Public Methods
-    
+
     func refreshBatteryInfo() {
         // Clear previous error
         errorMessage = nil
-        
+
         // Use the safe method which handles errors gracefully
         batteryInfo = reader.getBatteryInfoSafe()
-        
+
         // Check if we got default values (indicating an error)
         if batteryInfo.currentCharge == 0 && !batteryInfo.isCharging && !batteryInfo.isPluggedIn {
             // Try the throwing version to get the actual error
@@ -161,41 +188,43 @@ class BatteryViewModel: ObservableObject {
                 errorMessage = "Unable to read battery information"
                 logger.error("Unknown error: \(error)")
             }
-        } else {
-            logger.debug("Battery: \(self.batteryInfo.currentCharge)%, \(self.batteryInfo.isCharging ? "charging" : "not charging")")
+            return
         }
+
+        logger.debug("Battery: \(self.batteryInfo.currentCharge)%, \(self.batteryInfo.isCharging ? "charging" : "not charging")")
+
+        // Check for alert conditions and trigger notch glow
+        checkAndTriggerAlerts()
     }
-    
+
     // MARK: - Enhanced Notch Display Methods
-    
+
     var isNotchDisplayEnabled: Bool {
         notchViewModel.layoutMode != .off
     }
-    
+
     var notchLayoutMode: MicroverseNotchViewModel.NotchLayoutMode {
-        get { 
+        get {
             let currentMode = notchViewModel.layoutMode
             logger.debug("Getting notch layout mode: \(currentMode.displayName)")
-            return currentMode 
+            return currentMode
         }
-        set { 
+        set {
             let oldMode = notchViewModel.layoutMode
             logger.info("Setting notch layout mode from \(oldMode.displayName) to \(newValue.displayName)")
             notchViewModel.layoutMode = newValue
             saveSetting("notchLayoutMode", value: newValue.rawValue)
-            
-            // Update notch display based on new mode
+
             Task { @MainActor in
                 if newValue == .off {
                     try? await notchViewModel.hideNotch()
                 } else {
-                    // Show notch if switching from off, or refresh if already visible
                     try? await notchViewModel.showNotch()
                 }
             }
         }
     }
-    
+
     func toggleNotchDisplay() {
         Task { @MainActor in
             do {
@@ -210,14 +239,14 @@ class BatteryViewModel: ObservableObject {
             }
         }
     }
-    
+
     func setNotchDisplayEnabled(_ enabled: Bool) {
         if enabled {
             // Enable with current layout mode (default to split if off)
             if notchViewModel.layoutMode == .off {
                 notchLayoutMode = .split
             }
-            
+
             Task { @MainActor in
                 do {
                     try await notchViewModel.showNotch()
@@ -231,44 +260,117 @@ class BatteryViewModel: ObservableObject {
             notchLayoutMode = .off
         }
     }
-    
+
     var isNotchAvailable: Bool {
         NSScreen.main?.hasNotch ?? false
     }
-    
+
     var notchViewModelInstance: MicroverseNotchViewModel {
         notchViewModel
     }
-    
+
+    /// Triggers a test notch glow alert
+    func testNotchAlert(type: NotchAlertType) {
+        NotchGlowManager.shared.showAlert(type: type, duration: 2.0, pulseCount: 2)
+    }
+
+    // MARK: - Notch Alert Logic
+
+    private func checkAndTriggerAlerts() {
+        guard enableNotchAlerts else { return }
+        guard isNotchAvailable else { return }
+
+        let currentCharge = batteryInfo.currentCharge
+        let isCharging = batteryInfo.isCharging
+        let isPluggedIn = batteryInfo.isPluggedIn
+        let isOnBatteryPower = !isPluggedIn
+
+        // Defer storing previous state until after checks
+        defer {
+            previousBatteryCharge = currentCharge
+            previousIsCharging = isCharging
+            previousIsPluggedIn = isPluggedIn
+        }
+
+        // Skip first run (no previous data)
+        guard previousBatteryCharge >= 0 else { return }
+
+        // Alert: Charger connected (even if macOS reports not charging due to optimizations)
+        if isPluggedIn && !previousIsPluggedIn {
+            // Calm, deliberate ping-pong sweep for "charger connected".
+            NotchGlowManager.shared.showSuccess(duration: 4.0)
+            hasShownLowBatteryAlert = false
+            hasShownCriticalBatteryAlert = false
+            logger.info("Notch alert: Charger connected")
+            return
+        }
+
+        // Alert: Fully charged (reached 100% while plugged in)
+        if currentCharge >= 100 && previousBatteryCharge < 100 && isPluggedIn {
+            NotchGlowManager.shared.showSuccess(duration: 2.0)
+            logger.info("Notch alert: Fully charged")
+            return
+        }
+
+        // Alert: Critical battery (≤10%)
+        if currentCharge <= 10 && isOnBatteryPower && !hasShownCriticalBatteryAlert {
+            NotchGlowManager.shared.showCritical(duration: 3.0)
+            hasShownCriticalBatteryAlert = true
+            logger.info("Notch alert: Critical battery")
+            return
+        }
+
+        // Alert: Low battery (≤20%, crossed threshold)
+        if currentCharge <= 20 && previousBatteryCharge > 20 && isOnBatteryPower && !hasShownLowBatteryAlert {
+            NotchGlowManager.shared.showWarning(duration: 2.0)
+            hasShownLowBatteryAlert = true
+            logger.info("Notch alert: Low battery")
+            return
+        }
+
+        // Reset alert flags when battery recovers
+        if currentCharge > 20 {
+            hasShownLowBatteryAlert = false
+        }
+        if currentCharge > 10 {
+            hasShownCriticalBatteryAlert = false
+        }
+    }
+
     // MARK: - Private Methods
-    
+
     /// Starts battery monitoring with adaptive refresh rates
     /// - Slower refresh when battery is stable (plugged in at 100%)
     /// - Normal refresh when charging or on battery power
     /// - Faster refresh when battery is critically low
     private func startMonitoring() {
-        timer?.invalidate()
-        
-        // Calculate adaptive refresh interval based on battery state
-        let adaptiveInterval = calculateAdaptiveRefreshInterval()
-        
-        timer = Timer.scheduledTimer(withTimeInterval: adaptiveInterval, repeats: false) { [weak self] _ in
-            Task { @MainActor in
-                self?.refreshBatteryInfo()
-                // Schedule next update with potentially different interval
-                self?.startMonitoring()
+        batteryRefreshTask?.cancel()
+
+        batteryRefreshTask = Task { @MainActor [weak self] in
+            guard let self else { return }
+
+            while !Task.isCancelled {
+                let adaptiveInterval = self.calculateAdaptiveRefreshInterval()
+                self.logger.debug("Adaptive refresh: Next update in \(adaptiveInterval)s (battery: \(self.batteryInfo.currentCharge)%, charging: \(self.batteryInfo.isCharging))")
+
+                do {
+                    try await Task.sleep(nanoseconds: UInt64(adaptiveInterval * 1_000_000_000))
+                } catch {
+                    break
+                }
+
+                if Task.isCancelled { break }
+                self.refreshBatteryInfo()
             }
         }
-        
-        logger.debug("Adaptive refresh: Next update in \(adaptiveInterval)s (battery: \(self.batteryInfo.currentCharge)%, charging: \(self.batteryInfo.isCharging))")
     }
-    
+
     /// Calculates optimal refresh interval based on battery state
     /// Returns interval in seconds
     private func calculateAdaptiveRefreshInterval() -> TimeInterval {
         // Use user preference as base interval
         let baseInterval = refreshInterval
-        
+
         // Adaptive logic based on battery state
         if batteryInfo.currentCharge <= 5 {
             // Critical battery: Update every 2 seconds
@@ -290,67 +392,92 @@ class BatteryViewModel: ObservableObject {
             return baseInterval
         }
     }
-    
+
     private func setupBindings() {
         // All settings are now handled by didSet observers
     }
-    
+
     private func loadSettings() {
         let defaults = UserDefaults.standard
-        
+
         // Load all settings with defaults
         if defaults.object(forKey: "showPercentageInMenuBar") != nil {
             showPercentageInMenuBar = defaults.bool(forKey: "showPercentageInMenuBar")
         }
-        
+
         let savedInterval = defaults.double(forKey: "refreshInterval")
         if savedInterval > 0 {
             refreshInterval = savedInterval
         }
-        
+
         showDesktopWidget = defaults.bool(forKey: "showDesktopWidget")
-        
+
         if let styleRaw = defaults.string(forKey: "widgetStyle"),
            let style = WidgetStyle(rawValue: styleRaw) {
             widgetStyle = style
         }
-        
+
         // Load auto-update setting
         if defaults.object(forKey: "checkForUpdatesAutomatically") != nil {
             checkForUpdatesAutomatically = defaults.bool(forKey: "checkForUpdatesAutomatically")
         }
-        
+
         // Load notch layout mode
         if let modeRaw = defaults.string(forKey: "notchLayoutMode"),
            let mode = MicroverseNotchViewModel.NotchLayoutMode(rawValue: modeRaw) {
             notchViewModel.layoutMode = mode
         }
-        
+
+        // Load notch alerts setting (default true)
+        if defaults.object(forKey: "enableNotchAlerts") != nil {
+            enableNotchAlerts = defaults.bool(forKey: "enableNotchAlerts")
+        }
+
+        // Load startup notch animation setting (default true)
+        if defaults.object(forKey: "enableNotchStartupAnimation") != nil {
+            enableNotchStartupAnimation = defaults.bool(forKey: "enableNotchStartupAnimation")
+        }
+
         // Note: launchAtStartup is already loaded from LaunchAtStartup.isEnabled
     }
-    
+
     private func saveSetting(_ key: String, value: Any) {
         UserDefaults.standard.set(value, forKey: key)
     }
-    
+
+    private func playStartupNotchAnimationIfNeeded() async {
+        guard !hasPlayedStartupNotchAnimation else { return }
+        guard enableNotchAlerts, enableNotchStartupAnimation else { return }
+        guard isNotchAvailable else { return }
+
+        hasPlayedStartupNotchAnimation = true
+        await NotchGlowManager.shared.playStartupAnimation()
+    }
+
     // MARK: - Auto-Update Methods
-    
+
     private func schedulePeriodicUpdateCheck() {
         cancelPeriodicUpdateCheck()
-        
-        // Check every 24 hours
-        updateCheckTimer = Timer.scheduledTimer(withTimeInterval: 24 * 60 * 60, repeats: true) { _ in
-            Task { @MainActor in
+
+        updateCheckTask = Task { @MainActor in
+            while !Task.isCancelled {
+                do {
+                    try await Task.sleep(nanoseconds: 24 * 60 * 60 * 1_000_000_000)
+                } catch {
+                    break
+                }
+
+                if Task.isCancelled { break }
                 SecureUpdateService.shared.checkForUpdates()
             }
         }
-        
+
         logger.info("Scheduled periodic update checks every 24 hours")
     }
-    
+
     private func cancelPeriodicUpdateCheck() {
-        updateCheckTimer?.invalidate()
-        updateCheckTimer = nil
+        updateCheckTask?.cancel()
+        updateCheckTask = nil
         logger.info("Cancelled periodic update checks")
     }
 }
